move default_sim to core
