+++ packages/core/src/sims/common/potency_ratio.ts
