+++ packages/frontend/src/scripts/main.ts
@@ -1,7 +1,7 @@
-import {registerDefaultSims} from "./sims/default_sims";
 import {installDoubleClickHandler} from "@xivgear/common-ui/util/stop_double_click";
 import {processNav, processHashLegacy} from "./nav_hash";
 import {earlyUiSetup, initialLoad, initTopMenu} from "./base_ui";
+import { registerDefaultSims } from "@xivgear/core/sims/default_sims";
 
 // Main entry point for actual browsers
 document.addEventListener("DOMContentLoaded", () => {
