+++ packages/frontend/src/scripts/test/sims/multi_sim_run.ts
@@ -1,8 +1,8 @@
 import 'global-jsdom/register';
 import {HEADLESS_SHEET_PROVIDER} from "@xivgear/core/sheet";
 import {rangeInc} from "@xivgear/core/util/array_utils";
-import {registerDefaultSims} from "../../sims/default_sims";
 import { SgeSheetSim } from '@xivgear/core/sims/healer/sge_sheet_sim_mk2';
+import { registerDefaultSims } from '@xivgear/core/sims/default_sims';
 
 const sheetJson = '{"name":"7.05 SGE BiS","sets":[{"name":"INFORMATION-----------","items":{"Head":{"id":42840,"materia":[{"id":-1},{"id":-1}]}},"description":"","relicStatMemory":{}},{"name":"i727 vs i729","items":{"Weapon":{"id":42587,"materia":[{"id":-1},{"id":-1}]}},"description":"i727: Crafted Hands, Normal Boots. Generally provides the highest damage within its SpS bracket. \\n\\ni729: Crafted Hands, Savage Boots. Much less risky on your survivability at minimal cost to damage compared to i727, but does cost *some* damage.\\n\\nMelds will change slightly, but have been standardized to match each other and allow flexibility moving between ilvl sets AMAP.","relicStatMemory":{}},{"name":"Which GCD?","items":{"Weapon":{"id":42587,"materia":[{"id":-1},{"id":-1}]}},"description":"SGE can FINALLY run a clean 2.50 as it has always desired.\\n\\nHowever, we will include 2.49 and 2.48 in this sheet for those of you who want to run another healer that prefers those GCDs. Also for fun data comparatives, for all you nerds out there playing SGE like us.","relicStatMemory":{}},{"name":"--------i730---------","items":{"Head":{"id":42840,"materia":[{"id":-1},{"id":-1}]}},"description":"","relicStatMemory":{}},{"name":"Died of Phlegma | 2.50","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":43150,"materia":[{"id":41771},{"id":41773}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41773}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"Highest simming DPS, base GCD","relicStatMemory":{}},{"name":"Psyche-delic | 2.49","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":43150,"materia":[{"id":41771},{"id":41771}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41773}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41782},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"This set should really only be considered if you know your frame rate is on the lower end.","relicStatMemory":{}},{"name":"Kardia Arrest | 2.48","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":43150,"materia":[{"id":41773},{"id":41773}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41782},{"id":41782}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"This is a compat set that matches WHM\'s best simming set. If you want to flex between the two options and consider yourself way more comfortable on WHM, this set will best match what their strongest one is.","relicStatMemory":{}},{"name":"-------i727--------","items":{},"description":"","relicStatMemory":{}},{"name":"Infinite Rhizomata | 2.50","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41771},{"id":41771},{"id":41758},{"id":41758}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":42844,"materia":[{"id":41771},{"id":41771}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"Base GCD","relicStatMemory":{}},{"name":"Pneumarical Analysis | 2.49","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41771},{"id":41771},{"id":41769},{"id":41769}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":42844,"materia":[{"id":41771},{"id":41771}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"2x Quicktongue XI materia on Gloves","relicStatMemory":{}},{"name":"Haima name is SGE | 2.48","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41782},{"id":41782}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41771},{"id":41771},{"id":41758},{"id":41758}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":42844,"materia":[{"id":41771},{"id":41771}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"2x Quicktongue XII Materia on Chest","relicStatMemory":{}},{"name":"-------i729--------","items":{},"description":"","relicStatMemory":{}},{"name":"Wise Wolf Holos | 2.50","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41773},{"id":41771},{"id":41758},{"id":41758}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"0 SpS Melds","relicStatMemory":{}},{"name":"Over Dosis | 2.49","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41771},{"id":41771}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41773},{"id":41771},{"id":41769},{"id":41769}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"2x Quicktongue XI Materia on Gloves","relicStatMemory":{}},{"name":"Toxikon Personality | 2.48","items":{"Weapon":{"id":43119,"materia":[{"id":41772},{"id":41772}]},"Head":{"id":43071,"materia":[{"id":41772},{"id":41771}]},"Body":{"id":43149,"materia":[{"id":41782},{"id":41782}]},"Hand":{"id":42919,"materia":[{"id":41771},{"id":41773},{"id":41771},{"id":41758},{"id":41758}]},"Legs":{"id":43074,"materia":[{"id":41771},{"id":41771}]},"Feet":{"id":43152,"materia":[{"id":41772},{"id":41772}]},"Ears":{"id":43161,"materia":[{"id":41772},{"id":41771}]},"Neck":{"id":43089,"materia":[{"id":41771},{"id":41771}]},"Wrist":{"id":43171,"materia":[{"id":41771},{"id":41771}]},"RingLeft":{"id":43176,"materia":[{"id":41772},{"id":41771}]},"RingRight":{"id":43099,"materia":[{"id":41772},{"id":41772}]}},"food":44178,"description":"2x Quicktongue XII Materia on Chest","relicStatMemory":{}}],"level":100,"job":"SGE","partyBonus":5,"race":"Raen","sims":[{"stub":"sge-sheet-sim-mk2","settings":{"customSettings":{"usePotion":false,"includeInExport":true},"buffConfig":{"jobs":[],"buffs":["Mug","Battle Litany","Brotherhood","Arcane Circle","Searing Light","Embolden","Starry Muse","Technical Finish","Battle Voice","Radiant Finale","Chain","Divination"]},"cycleSettings":{"cycles":6,"totalTime":720,"which":"totalTime","useAutos":false},"resultSettings":{"stdDevs":0}},"name":"SGE Sim Mk.II"}],"itemDisplaySettings":{"minILvl":680,"maxILvl":999,"minILvlFood":640,"maxILvlFood":999,"higherRelics":true},"mfni":false,"mfp":["spellspeed","crit","dhit","determination","piety"],"mfMinGcd":2.05,"customItems":[],"customFoods":[]}';
